plugins/udevng: specially handle sub-tty devices under USB subsystem
check_subtty_device() performs a special detection of a sub-TTY device,
which is a child of an existing USB device. /dev/ttyACM0 made by an
USB-ethernet interface is a good example. Unfortunately such devices
cannot be uniquely determined under the USB subsystem, because udev
recognizes the devices as a pair of bogus vendor ID and model ID,
given from a individual USB host controller. Thus we need to do a
manual udev enumeration again, to be able to detect the TTY device
that resides under a given parent USB device.

For example, when syspath is
"/devices/pci0000:00/0000:00:14.0/usb3/3-1/3-1:1.2/tty/ttyACM0",
parent_syspath would be
"/devices/pci0000:00/0000:00:14.0/usb3/3-1/3-1:1.2".

With this change, oFono can detect /dev/ttyACM* devices as expected.