ubloxmodem: push back APN into gprs context
It can be that the APN is different than what a user provided. Especially
when using the cid activated from a GCEV callback. A user tries to enable
another APN, it will fail and oFono will fall back to the active context
received via ME PDN ACT CGEV event. This is quite normal for a 4G network.