ublox: tobyl2: support more internet contexts
Toby L2 supports 8 defined PDP context and 3 acvite ones. This patch
creates the same number of gprs-context structures as supported PDP
contexts. Every context activation will select an unused gprs-context
instance.

This patch also deals with the subtleties of having a default EPS
bearer/PDP context:

* If a context activation event is received without havning any
previously activated contexts, take that as being the default EPS
bearer. On next activation use it if that activation fails.

* Re-read the APN for only for a default EPS bearer. Sometimes,
after activation, the APN changes to a value that is unusable in a
subsequent activation. For example:
web.vodafone.de -> events.vodafone.de.mcc.mng.gprs
For a default EPS bearer we don't care, because we don't have
control over the APN. But we do care for other contexts because
we want to re-activate them more than once.