drivers/ubloxmodem: support also the router mode
In addition to the bridge mode, U-Blox TOBY-L2 should also support the
router mode, which has been the default mode in the previous models.

Unfortunately, on the U-Blox TOBY-L2, its embedded DHCP server does not
work as expected. So it's not possible to run a DHCP client for the IPv4
networking. To solve that, here we configure a static IP address.
While IP addresses, netmask, gateway addresses could be fetched via
+UIPCONF, its DNS server and APN cannot be fetched with that command.
So an additional AT command, +CGCONTRDP should be used again.