drivers/ubloxmodem: add support for the secondary PDP context
Introduce a new feature for the secondary PDP context.
A secondary PDP context can be optionally enabled for several models
like U-Blox TOBY-L2. Its context ID is different from the primary
context ID, but its APN must be a substring of the primary APN.

So the approach used here like this:
When activating a primary context, it tries to determine if the current
context contains a possible APN that can be used as secondary context.
This is done by looking up every APN in the global context list.
After activating a context, if the current context is secondary,
send a "+CGDSCONT" command, to set a pair of activated contexts,
(primary,secondary).

NOTE: this feature is highly experimental, but unfortunately it cannot
be tested, because ordinary SIMs don't support activation of such
secondary contexts. So this part should be fully tested, once test
hardware would be available in the future.

Change-Id: I8e0c8b331905e7479f0244c413996fe33b97a593